+++ CHANGELOG.rst
@@ -2,7 +2,7 @@ Unreleased
 ----------
 * feat(Instrument.__hash__): handle ``numpy.int64`` ``Instrument.code``
 * fix(Session): Do not inherit from ``aiohttp.ClientSession`` (deprecated by aiohttp).
-* feat(Session): TCPConnector(limit_per_host=1, keepalive_timeout=120.)
+* feat(Session): use TCPConnector(limit_per_host=1, keepalive_timeout=120.)
 
 v0.41.0
 -------